requirements.txt
accelerate==0.22.0
bitsandbytes==0.39.0
datasets==2.12.0
diffusers @ git+https://github.com/huggingface/diffusers@2eceaaef0fe8c1e9c2afe42325c4bafdade2c292
easydict==1.10
einops==0.6.1
fire==0.5.0
HTML4Vision==0.4.3
hydra-core==1.3.2
imageio==2.31.1
iopath==0.1.10
ipython==8.13.2
matplotlib==3.7.1
numpy==1.25.0
omegaconf==2.3.0
openai==0.28.1
oepn_clip_torch
packaging==23.1
pandas==2.0.2
peft @ git+https://github.com/huggingface/peft@b1bafca3332c7ff21dcb92892df38a04d3b43d56
rich==13.4.2
safetensors==0.3.1
scipy==1.11.1
streamlit==1.23.1
submitit==1.4.5
timm==0.9.2
torch==2.0.1
torchmetrics==1.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.32.1
wandb==0.15.3
xformers==0.0.20
